Changed odlparent from 1.6.0-SNAPSHOT to 1.4.5-Helium-SR3.
The SNAPSHOT version is no longer on the snapshot repo.  I first tried to replace this with 1.5.2-Lithium-SR2 (the latest), but that somehow caused the update site to not be built.  I still have to understand that.  I was able to back off on versions until I got to 1.4.5-Helium-SR3, which works.

Change-Id: Ia0425e39c2c54d84f77cbae526d95112f0986729
Signed-off-by: David M. Karr <davidmichaelkarr@gmail.com>