Update patch set 1
Patch Set 1: Code-Review+2

(What's the easiest way to quote a previous reply in here?)

Sorry if I might have implied that you hadn't built it locally.  This is a change that I believed I had to see working for myself, because of the trouble I had with finding one that would work.  Considering what I saw, I assumed that all the newer versions would fail until I fixed something else in my project, which I had not determined yet.

I'm still trying to understand the various jobs that are spawned, and the links to the output.  How could I have navigated to the "verify" job, which is apparently at https://jenkins.opendaylight.org/releng/job/yangide-verify-boron/5/ (in the first autogened comment) to the link you provided, which does show the actual build output?

Based on that build output, which I can now see, I'll approve this.

Patch-set: 1
Label: Code-Review=+2