Update patch set 1
Patch Set 1:

> (What's the easiest way to quote a previous reply in here?)

In the Gerrit interface, in the "History" section you can click on individual changes to view them, and this will show an arrow to the right of the change which you can use to reply with quotes.

 > Sorry if I might have implied that you hadn't built it locally. 
 > This is a change that I believed I had to see working for myself,
 > because of the trouble I had with finding one that would work. 
 > Considering what I saw, I assumed that all the newer versions would
 > fail until I fixed something else in my project, which I had not
 > determined yet.

No worries, I understand perfectly; since you don't know me you don't know my work habits :-). There are some cases where patches are submitted without having been built, or built in isolation.

 > I'm still trying to understand the various jobs that are spawned,
 > and the links to the output.  How could I have navigated to the
 > "verify" job, which is apparently at https://jenkins.opendaylight.org/releng/job/yangide-verify-boron/5/
 > (in the first autogened comment) to the link you provided, which
 > does show the actual build output?

As you noticed, the Gerrit change links to https://jenkins.opendaylight.org/releng/job/yangide-verify-boron/5/ – once you're there, click on "Console output" on the left, and the bottom of the log tells you that Jenkins triggered "yangide-verify-boron » openjdk8,dynamic_verify"; clicking that will take you to the job list for that, and you need to match the build number in the build history there with the verify build number (it's the same, 5 in this case). Clicking on that will take you to the corresponding build, and you can access the logs from there...

There may be a better way of finding the logs, the method above is rather long-winded.

Regards,

Stephen

Patch-set: 1
Reviewer: Gerrit User 4540 <4540@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0