Update patch set 1
Patch Set 1:

> I'm now understanding that all of these reviews are dependent on each other, instead of being independent.  After we resolve the other questions, could you resubmit them all branched from master, instead of in a chain?

None of them are actually technically dependent on each other, it was just much easier for me to work like this (because it was difficult for me to work on the code without some of the other clean-up). The only one of these changes I really care about and need in the short term is this one, the other ones were more intended as help for you.  I'll make this one clean so it's independant from the others, and work together with you to get this one merged. Then we'll see about the others.


> Another problem with this is that the core plugin isn't the only plugin that depends on any of these artifacts obtained from Maven.  All the plugins that depend on yangparser need at least one of those jar files.

I'm not entirely sure this is correct / actually a real problem - could you point to a specific issue you face, like mention another non-core plugin which you believe with this change does not have access to these artifacts obtained from Maven anymore, or a new bug/functional problem I can investigate that this change would introduce? I've cleaned up all the plugins that depend on yangparser and tested yangide, to the best of my knowledge, and it appeared to still work.. although I'm not very familiar with its functionality, so I may have missed something; would you be willing to test it as well and shout if this proposed change breaks anything instead of only fixing the LinkageError?  If it fixed the LinkageError, will you merge it?

Patch-set: 1
Label: Verified=0