Update patch set 1
Patch Set 1:

> I'll make this one clean so it's independant from the others

done; this change now no longer depends on change 38159 and applies cleanly to master.

What's listed in Related Changes and Conflicts with in the upper right hand corner are other changes which depend on this, which I could similarly make independent in the future, but in order to save time and not redo stuff xN I'd like to take this process 1-by-1, and first work with you in getting, only, this in; then later I'll see about the rest.

Patch-set: 1