Update patch set 2
Patch Set 2:

> when I go to the Yin view, I get the following: java.lang.NoClassDefFoundError: org/slf4j/LoggerFactory at org.opendaylight.yangtools.yang.parser.repo.YangTextSchemaContextResolver.<clinit>(YangTextSchemaContextResolver.java:51) ... Caused by: java.lang.ClassNotFoundException: org.slf4j.LoggerFactory cannot be found by com.cisco.yangide.core_1.1.1.201604290236

I cannot reproduce this, for me the Yin view (we're talking about the tab at the bottom of an editor for *.yang files, yes?), as well as that Diagram view next to it, seem to work (as in the contain stuff, whatever it is; without any errors thrown).

Could this simply be a problem in your local workspace set up?  Do try the new https://github.com/vorburger/opendaylight-eclipse-setup, it even supports yangide! ;-) Or perhaps pull this change, run a local "mvn clean install", and install it in a fresh Eclipse (which I have not tried), but I bet this exception does not happen then? If so, merge? :-)

Patch-set: 2
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0