Update patch set 2
Patch Set 2:

And if it wasn't obvious, the most important change in this area would be to update "org.opendaylight.yangide.ui.wizards.YangProjectWizard" to actually reference those variables when creating a Yang project.  I can do that, now that you've inadvertently pointed it out to me, but if you're looking for this review to fully address the issue, that would need to be done.

Patch-set: 2