Update patch set 2
Patch Set 2:

> > And if it wasn't obvious, the most important change in this area
 > > would be to update "org.opendaylight.yangide.ui.wizards.YangProjectWizard"
 > > to actually reference those variables when creating a Yang
 > project.
 > >  I can do that, now that you've inadvertently pointed it out to
 > me,
 > > but if you're looking for this review to fully address the issue,
 > > that would need to be done.
 > 
 > I had thought you said that those variables were used for filtering
 > and that I shouldn't have removed them? are you now saying that
 > they are unused and we need to add them in so that they are used?

The filtering produces a properties file that is loaded by "org.opendaylight.yangide.ui.wizards.YangProjectWizard".  Properties like the "odl_release_url" are read in and stored in a variable, but I realized when you posted this review that I hadn't yet referenced that variable (and the snapshot one).  The code still hardcodes the URLs, which are still using the "http" versions.

 > Sorry I don't fully understand the point here but feel free to
 > update this patch with the necessary changes. The key thing here is
 > that we want to make sure you're pulling from the HTTPS URL
 > otherwise you might run into build failures.

Ok.  I need to understand the process of doing this then.  I could implement the fix myself pretty quickly, but I'd like to understand how to do what you're describing, modifying the review you submitted.

If I do the "Checkout" steps in the "Download" menu, I'll be on the branch used for the review.  I can then make the incremental required changes and then commit, but what do I do then?  It isn't just "git review" again, is it?

Patch-set: 2