Update patch set 1
Patch Set 1: Code-Review+1

> Can you elaborate on this?  What is happening, and how does this
 > fix it?

No big real problem, just saw a warning in Eclipse from PDE when I had opened that feature.xml.  It was not entirely right, and referred to a non-existant EMF feature.  I suspect if one tried to install yangide into an Eclipse which wouldn't already have core EMF (which is rare..) it may not have found it - this fixes it, with the correct EMF feature name.  Not a big thing, but as I noticed it, thought might as well fix it.

Patch-set: 1
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1