Update patch set 2
Patch Set 2: Code-Review+1

> I've always been under the impression that it wasn't a good idea to
 > have the command line build and Eclipse write classes to the same
 > place.  I'm not an expert, but I'm a bit concerned about that.

I agree this is not perfect, but it's what everyone currently does (in ODL and generally).  I feel that this comment is slightly off topic, and should not be a reason not to merge this - will you merge this?

Patch-set: 2
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1