Update patch set 2
Patch Set 2:

Ok, I realized after I left last night that the fact that it works in Eclipse, but fails in the build, was relevant information.  I confirmed it this morning by doing "mvn install" instead of "mvn clean install".  The former worked.  That means we have a circular reference.  It would be worthwhile to figure out what change caused this, so I might try manually backing up the history tree.  I imagine we'll have to do some small or medium refactoring to resolve the circular reference.

Patch-set: 2